test/integration/gles/replay: Split tests into multiple files
